\name{BiomassEst.Info}
\alias{BiomassEst.Info}
\title{Package Information}
\usage{
BiomassEst.Info(topic)
}
\description{
Prints out information about the BiomassEST package.
}
\arguments{
\item{topic}{
If not set, general information about the package will be printed. Use 'Species' in order to return a data.frame with all the currently supported species.
}
}
