\name{Corrections}
\alias{Corrections}
\title{
Measurement height corrections
}
\description{
Correction values for different heights (cm) of the stem diameter measurements
}
\references{
Anninghoefer et al., 2016. Species-specific and generic biomass equations for seedlings and saplings of European tree species. Eur J Forest Res 135 2016, pp 313–329. doi: 10.1007/s10342-016-0937-z
}
